+++ src/image.rs
@@ -241,10 +241,10 @@ impl Image {
     }
 
     /// The indicated bit_depth corresponds to the bit depth per channel.
-    /// I.e. for interleaved formats like RRGGBB, the bit_depth would be, e.g., 10 bit instead
-    /// of 30 bits or 3*16=48 bits.
-    /// For backward compatibility, one can also specify 24bits for RGB and 32bits for RGBA,
-    /// instead of the preferred 8 bits.
+    /// I.e. for interleaved formats like RRGGBB, the bit_depth would be,
+    /// e.g., 10 bit instead of 30 bits or 3*16=48 bits.
+    /// For backward compatibility, one can also specify 24bits for RGB and
+    /// 32bits for RGBA, instead of the preferred 8 bits.
     pub fn create_plane(
         &mut self,
         channel: Channel,